feat : replace namespace Core to Common
As it is more proper to the purpose of source codes and may confuse with Core directory in CubbyFlow, I decided to rename Core namespace to Common
