+++ Sources/GL3/DebugUtils.cpp
@@ -1,4 +1,4 @@
-#include <Core/Macros.hpp>
+#include <Common/Macros.hpp>
 #include <GL3/DebugUtils.hpp>
 #include <cstdio>
 #include <iostream>
